requirements.txt
langchain
sentence-transformers
faiss-cpu
openai
python-docx
PyPDF2
scikit-learn